requirements.txt
fastapi
uvicorn[standard]
python-multipart
ultralytics
opencv-python-headless
numpy
pillow
streamlit
requests

